pyproject.toml
[project]
name = "flexdata-metric-prediction"
version = "0.1.0"
description = "Learned cost model and metric prediction for SQL workloads."
authors = [ { name = "Shaswat Gupta", email = "shaswat.gupta@ibm.com" }]
requires-python = ">=3.12,<4"
readme = "README.md"
license = "Apache-2.0"
dependencies = [
    "PyYAML>=6,<7",
    "graphviz>=0.21,<0.22",
    "numpy>=2.1,<3",
    "torch>=2.1,<3",
    "matplotlib>=3.8,<4",
    "scikit-learn>=1.7,<2",
    "pandas>=2.3,<3",
    "torch-geometric>=2.6,<3",
    "rich>=14,<15",
    "substrait>=0.24,<0.25",
    "wandb>=0.17",
    "seaborn>=0.13.2",
    "scipy>=1.16.3",
]

[project.optional-dependencies]
clearml = ["clearml==2.0.2"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0,<5",
    "git-lint>=0.1.2,<0.2",
    "detect-secrets",
    "black>=25.1.0,<26",
    "ruff>=0.14.0,<0.15",
    "mypy==1.18.2",
]
test = [
    "pytest==9.0.1",
    "pytest-cov==7.0.0",
]

[tool.uv]
default-groups = [
    "dev",
    "test",
]


[tool.uv.sources]
detect-secrets = { git = "https://github.com/ibm/detect-secrets.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ensure that the library is available to the tests
[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
